Move shared table events out into shared macro
