+++ uncharted_phoenix/lib/uncharted_phoenix/components/live_bar.ex
@@ -4,10 +4,7 @@ defmodule UnchartedPhoenix.LiveBarComponent do
   """
 
   use Phoenix.LiveComponent
-
-  def mount(socket) do
-    {:ok, assign(socket, :show_table, false)}
-  end
+  use UnchartedPhoenix.TableEvents
 
   def update(assigns, socket) do
     x_axis = assigns.chart.dataset.axes.magnitude_axis
@@ -35,13 +32,5 @@ defmodule UnchartedPhoenix.LiveBarComponent do
 
   def render(assigns) do
     Phoenix.View.render(UnchartedPhoenix.ComponentView, "live_bar.html", assigns)
-  end
-
-  def handle_event("show_table", _, socket) do
-    {:noreply, assign(socket, :show_table, true)}
-  end
-
-  def handle_event("hide_table", _, socket) do
-    {:noreply, assign(socket, :show_table, false)}
   end
 end
