+++ uncharted_phoenix/lib/uncharted_phoenix/components/live_pie.ex
@@ -4,10 +4,7 @@ defmodule UnchartedPhoenix.LivePieComponent do
   """
 
   use Phoenix.LiveComponent
-
-  def mount(socket) do
-    {:ok, assign(socket, :show_table, false)}
-  end
+  use UnchartedPhoenix.TableEvents
 
   def update(assigns, socket) do
     socket =
@@ -21,13 +18,5 @@ defmodule UnchartedPhoenix.LivePieComponent do
 
   def render(assigns) do
     Phoenix.View.render(UnchartedPhoenix.ComponentView, "live_pie.html", assigns)
-  end
-
-  def handle_event("show_table", _, socket) do
-    {:noreply, assign(socket, :show_table, true)}
-  end
-
-  def handle_event("hide_table", _, socket) do
-    {:noreply, assign(socket, :show_table, false)}
   end
 end
