+++ uncharted_phoenix/lib/uncharted_phoenix/components/live_progress.ex
@@ -5,10 +5,7 @@ defmodule UnchartedPhoenix.LiveProgressComponent do
 
   alias Uncharted.ProgressChart
   use Phoenix.LiveComponent
-
-  def mount(socket) do
-    {:ok, assign(socket, :show_table, false)}
-  end
+  use UnchartedPhoenix.TableEvents
 
   def update(assigns, socket) do
     socket =
@@ -23,13 +20,5 @@ defmodule UnchartedPhoenix.LiveProgressComponent do
 
   def render(assigns) do
     Phoenix.View.render(UnchartedPhoenix.ComponentView, "live_progress.html", assigns)
-  end
-
-  def handle_event("show_table", _, socket) do
-    {:noreply, assign(socket, :show_table, true)}
-  end
-
-  def handle_event("hide_table", _, socket) do
-    {:noreply, assign(socket, :show_table, false)}
   end
 end
